revised tests
